// src/models/pokemon.interface.ts
export interface Pokemon {
    name?: string;
    num?: number;
    image?: string;
}